pyproject.toml
[tool.poetry]
name = "automl"
version = "0.1.0"
description = ""
authors = ["Matheus Oliveira de Souza <msouza.os@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.db.dependencies]
sqlalchemy = "^2.0.23"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.0"
pytest = "^7.4.3"


[tool.poetry.group.api.dependencies]
fastapi = {extras = ["uvicorn"], version = "^0.105.0"}
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-multipart = "^0.0.6"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.2"
httpx = "^0.26.0"


[tool.poetry.group.ml.dependencies]
pandas = "^2.1.4"
chardet = "^5.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
